+++ src/biodwh2-core/src/main/java/de/unibi/agbi/biodwh2/core/io/mvstore/MVStoreCollection.java
@@ -125,9 +125,11 @@ public final class MVStoreCollection<T extends MVStoreModel> implements Iterable
         if (oldModel != null)
             for (final String key : oldModel.keySet()) {
                 final MVStoreIndex index = indices.get(key);
-                final Object property = oldModel.get(key);
-                if (index != null && property != null)
-                    index.remove(property, oldModel.getId());
+                if (index != null) {
+                    final Object property = oldModel.get(key);
+                    if (property != null)
+                        index.remove(property, oldModel.getId());
+                }
             }
     }