+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Released:
  * ```[feature]``` IO format readers now use Charset class instead of charset name
  * ```[feature]``` OBO ontologies now share a unified graph exporter
  * ```[feature]``` CSV and TSV format reader in FileUtils now ignores missing columns
+ * ```[fix]``` Fix missing ids in node mapping layer
  * ```[fix]``` Fix HTTP FTP pre web source parsing
  * ```[docs]``` Data source modules are now listed as usable instead of done, as always something might change
  * ```[docs]``` Updated development docs with new IntelliJ Idea version